Split the artifact-explorer's specs out of its sources
Moves every spec and the oracle and arena harnesses under `tests/`, leaving `src/`
as shipped code only, and lifts `loot.json` into `data/`. Developers can tell at a
glance which files are the product and which only measure it, and the test tree is
selectable as a unit by path.
