+++ wasmegg/artifact-explorer/data/.ignore
@@ -1,0 +1,1 @@
+*.json
