+++ wasmegg/artifact-explorer/package.json
@@ -35,15 +35,15 @@
     "build": "vue-tsc --noEmit && vite build",
     "dev": "vite",
     "fastbuild": "vite build",
-    "lint": "eslint --ext .js,.ts,.vue src",
+    "lint": "eslint --ext .js,.ts,.vue src tests",
     "serve": "vite preview",
-    "format": "prettier --write \"src/**/*.{js,ts,vue}\"",
+    "format": "prettier --write \"{src,tests}/**/*.{js,ts,vue}\"",
     "test": "vitest run",
-    "test:oracle": "RUN_ORACLE=1 vitest run src/oracle/optimizer-oracle.spec.ts --disable-console-intercept",
-    "arena": "ARENA=${ARENA:-smoke} vitest run src/oracle/arena/invariants.spec.ts --disable-console-intercept",
-    "arena:check": "vitest run src/oracle/arena/independence.spec.ts",
+    "test:oracle": "RUN_ORACLE=1 vitest run tests/oracle/optimizer-oracle.spec.ts --disable-console-intercept",
+    "arena": "ARENA=${ARENA:-smoke} vitest run tests/arena/invariants.spec.ts --disable-console-intercept",
+    "arena:check": "vitest run tests/arena/independence.spec.ts",
     "coverage": "vitest run --coverage",
     "test:watch": "vitest",
-    "repro": "vite-node -c vitest.config.ts src/oracle/repro.ts --"
+    "repro": "vite-node -c vitest.config.ts tests/oracle/repro.ts --"
   }
 }
