+++ wasmegg/artifact-explorer/src/lib/solver/SPEC.md
@@ -1,9 +1,9 @@
 # The plan as a mixed-integer program
 
-The shipped planner. `src/lib/optimizer-core.ts` calls `solveWith` from `oa.ts` on a
+The shipped planner. `optimizer-core.ts` calls `solveWith` from `oa.ts` on a
 module loaded by `loadHighs`, and the arena enters the same pair through a shim
-(`src/oracle/arena/solvers/highs/`), so the solver users run and the solver the
-harness measures are one code path. `src/oracle/arena/ARENA.md` states the problem;
+(`tests/arena/solvers/highs/`), so the solver users run and the solver the
+harness measures are one code path. `tests/arena/ARENA.md` states the problem;
 this file states the method.
 
 The whole problem — mission counts per slot, crafts as flow over the conservation
@@ -219,7 +219,7 @@ MILP's own columns — the packing witness that summing over slots threw away.
 
 `SLOT_TOL` is 1e-9 not for resolution — the drift it absorbs is three decades
 smaller — but because that is the judge's own packing tolerance
-(`src/oracle/arena/pack-feasibility.ts`). It is a ceiling rather than a preference:
+(`tests/arena/pack-feasibility.ts`). It is a ceiling rather than a preference:
 anything looser certifies plans the judge calls infeasible, an arena C1 hard
 failure.
