+++ wasmegg/artifact-explorer/tests/arena/contract.ts
