+++ wasmegg/artifact-explorer/tests/arena/invariants.spec.ts
