+++ wasmegg/artifact-explorer/tests/arena/pack-feasibility.spec.ts
