+++ wasmegg/artifact-explorer/tests/arena/pack-feasibility.ts
