+++ wasmegg/artifact-explorer/tests/arena/registry.ts
