+++ wasmegg/artifact-explorer/tests/unit/optimizer-cost.spec.ts
@@ -8,10 +8,10 @@ import {
   fractionalCraftCost,
   previousCraftsOf,
   sumCraftChainCost,
-} from './optimizer-cost';
-import { computeCraftChainTree } from './optimizer-tree';
+} from '@/lib/optimizer-cost';
+import { computeCraftChainTree } from '@/lib/optimizer-tree';
 import { lt1, lt2, lt3, lt4, makeNode, makeSolution, totemDag } from './spec-helpers';
-import type { RecipeDAG } from './types';
+import type { RecipeDAG } from '@/lib/types';
 
 const Name = ei.ArtifactSpec.Name;
 const Level = ei.ArtifactSpec.Level;
